chore: refactor class layout
