+++ src/Uno.Sdk/Services/NuGetClient.cs
@@ -6,8 +6,9 @@ using System.Net.Http;
 using System.Net.Http.Json;
 using System.Text;
 using System.Threading.Tasks;
+using Uno.Sdk.Models;
 
-namespace Uno.Sdk;
+namespace Uno.Sdk.Services;
 
 #nullable enable
 public class NuGetClient : IDisposable
